--- memento/src/main/java/com/memento/server/api/controller/post/dto/read/PostAuthor.java
@@ -1,13 +1,0 @@
-package com.memento.server.api.controller.post.dto.read;
-
-import java.time.LocalDateTime;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.experimental.SuperBuilder;
-
-@Getter
-@SuperBuilder
-@AllArgsConstructor
-public class PostAuthor extends Author{
-}
